// demo-mcp-21-code-generator-node/input/reference-repos/ECOMMERCE_V9/MSPrChValGenerate/src/data-provider/model/http/request-config-http.model.ts
/**
 * Se define la configuracion de las interfaces usadas para consumos a legados
 * @author Alexis Sterzer
 */

export interface IRequestConfigHttp {
    method: EHttpMethod;
    url: string
    headers?: any;
    params?: any;
    data?: any;
    preambleCRLF?: boolean;
    postambleCRLF?: boolean;
}


export enum EHttpMethod {
    get = 'GET',
    delete = 'DELETE',
    post = 'POST',
    put = 'PUT',
}


export interface IRequestConfigHttpSOAP {
    url: string
    soapAction: any;
    data: string;
}